Add sqlx package; initial DB creation script; moving things between modules
