+++ Cargo.toml
@@ -16,5 +16,17 @@ actix-web = "4"
 serde = { version = "1.0.158", features = ["derive"] }
 tokio = { version = "1", features = ["macros", "rt-multi-thread"] }
 
+[dependencies.sqlx]
+version = "0.6.2"
+default-features = false
+features = [
+    "runtime-tokio-rustls",
+    "macros",
+    "postgres",
+    "uuid",
+    "chrono",
+    "migrate"
+]
+
 [dev-dependencies]
 reqwest = "0.11.14"
