+++ bin/init_db.sh
@@ -1,0 +1,54 @@
+#!/usr/bin/env bash
+set -x
+set -eo pipefail
+
+if ! [ -x "$(command -v psql)" ]; then
+  echo >&2 'Error: psql is not installed.'
+  exit 1
+fi
+
+if ! [ -x "$(command -v sqlx)" ]; then
+  echo >&2 'Error: sqlx is not installed.'
+  echo >&2 'Install it with: cargo install sqlx-cli --no-default-features --features rustls,postgres'
+  exit 1
+fi
+
+DB_USER=${POSTGRES_USER:=postgres}
+DB_PASSWORD=${POSTGRES_PASSWORD:=password}
+DB_NAME=${POSTGRES_DB:=newsletter}
+DB_PORT=${POSTGRES_PORT:=5432}
+DB_HOST=${POSTGRES_HOST:=localhost}
+
+if [[ -z "${SKIP_DOCKER}" ]]
+then
+  docker run \
+      -e POSTGRES_USER=${DB_USER} \
+      -e POSTGRES_PASSWORD=${DB_PASSWORD} \
+      -e POSTGRES_DB=${DB_NAME} \
+      -p ${DB_PORT}:5432 \
+      -d postgres \
+      postgres -N 1000
+fi
+
+export PGPASSWORD="${DB_PASSWORD}"
+until psql -h "${DB_HOST}" -U "${DB_USER}" -p "${DB_PORT}" -d "postgres" -c '\q'; do
+  >&2 echo "Postgres is still unavailable - sleeping"
+  sleep 1
+done
+
+>&2 echo "Postgres is up and running on port ${DB_PORT}!"
+
+
+DATABASE_URL=postgres://${DB_USER}:${DB_PASSWORD}@${DB_HOST}:${DB_PORT}/${DB_NAME}
+
+export DATABASE_URL
+
+>&2 echo "Creating database ${DB_NAME}..."
+
+sqlx database create
+
+>&2 echo "Running migrations..."
+
+sqlx migrate run
+
+>&2 echo "Database ${DB_NAME} is ready!"
