+++ src/main.rs
@@ -1,4 +1,4 @@
-use email_newsletter::run;
+use email_newsletter::startup::run;
 use std::net::TcpListener;
 
 #[tokio::main]
